pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ennam-django-apidog"
version = "0.1.0"
description = "Django integration for APIDOG - Export, sync, and manage OpenAPI schemas"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Ennam", email = "ennam@example.com"}
]
maintainers = [
    {name = "Ennam", email = "ennam@example.com"}
]
requires-python = ">=3.8"
keywords = ["django", "apidog", "openapi", "swagger", "drf", "rest", "api", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Django>=3.2",
    "djangorestframework>=3.12",
    "drf-spectacular>=0.26",
    "requests>=2.25",
    "PyYAML>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "responses>=0.24.0",
    "pre-commit>=3.5.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "django-stubs>=4.2.6",
    "djangorestframework-stubs>=3.14.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "build>=1.0",
    "twine>=4.0",
]

[project.urls]
Homepage = "https://github.com/ennam/ennam-django-apidog"
Documentation = "https://github.com/ennam/ennam-django-apidog#readme"
Repository = "https://github.com/ennam/ennam-django-apidog"
Changelog = "https://github.com/ennam/ennam-django-apidog/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/ennam/ennam-django-apidog/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ennam_django_apidog = [
    "templates/*",
    "py.typed",
]

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "test_*.py"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "ennam_django_apidog.*"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
